separate tests because of shared system property
