+++ zookeeper-server/src/test/java/org/apache/zookeeper/server/quorum/EphemeralNodeThrottlingTest.java
@@ -24,7 +24,8 @@ import org.apache.zookeeper.KeeperException;
 import org.apache.zookeeper.ZooDefs;
 import org.apache.zookeeper.ZooKeeper;
 import org.apache.zookeeper.metrics.MetricsUtils;
-import org.junit.Test;
+import org.apache.zookeeper.server.ServerMetrics;
+import org.junit.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -45,15 +46,33 @@ public class EphemeralNodeThrottlingTest extends QuorumPeerTestBase {
     static final int NUM_SERVERS = 5;
     static final String TEST_PATH = "/ephemeral-throttling-test";
 
-    @Test
-    public void byteSizeTest() throws Exception {
-        int totalEphemeralNodesByteLimit = 200;
-        System.setProperty(EPHEMERAL_BYTE_LIMIT_KEY, Integer.toString(totalEphemeralNodesByteLimit));
+    @BeforeClass
+    public static void setUpClass() {
+        System.setProperty(EPHEMERAL_BYTE_LIMIT_KEY, Integer.toString(DEFAULT_EPHEMERALNODES_TOTAL_BYTE_LIMIT));
+    }
+
+    @AfterClass
+    public static void tearDownClass() {
+        System.clearProperty(EPHEMERAL_BYTE_LIMIT_KEY);
+    }
+
+    @Before
+    public void setUpMethod() throws Exception {
         servers = LaunchServers(NUM_SERVERS);
+    }
+
+    @After
+    public void tearDownMethod() throws Exception {
+        servers.shutDownAllServers();
+    }
+
+    @Test
+    public void byteSizeTest() {
+        System.setProperty(EPHEMERAL_BYTE_LIMIT_KEY, Integer.toString(DEFAULT_EPHEMERALNODES_TOTAL_BYTE_LIMIT));
         ZooKeeper leaderServer = servers.zk[servers.findLeader()];
         int cumulativeBytes = 0;
         int i = 0;
-        while (cumulativeBytes <= totalEphemeralNodesByteLimit) {
+        while (cumulativeBytes <= DEFAULT_EPHEMERALNODES_TOTAL_BYTE_LIMIT) {
             cumulativeBytes += BinaryOutputArchive.getSerializedStringByteSize(TEST_PATH +i);
             try {
                 leaderServer.create(TEST_PATH + i++, new byte[512], ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.EPHEMERAL);
@@ -62,15 +81,17 @@ public class EphemeralNodeThrottlingTest extends QuorumPeerTestBase {
             }
         }
         long actual = (long) MetricsUtils.currentServerMetrics().get(EPHEMERAL_BYTE_LIMIT_VIOLATION_KEY);
+        System.out.println("byte limit property is: " + Integer.getInteger(EPHEMERAL_BYTE_LIMIT_KEY));
         assertEquals(1, actual);
 
-        servers.shutDownAllServers();
+
+        // waitForAll(servers, ZooKeeper.States.CONNECTING);
+        // ServerMetrics.DEFAULT_METRICS_FOR_TESTS.resetAll();
     }
 
     @Test
     public void limitingEphemeralsTest() throws Exception {
         System.setProperty(EPHEMERAL_BYTE_LIMIT_KEY, Integer.toString(DEFAULT_EPHEMERALNODES_TOTAL_BYTE_LIMIT));
-        servers = LaunchServers(NUM_SERVERS);
         ZooKeeper leaderServer = servers.zk[servers.findLeader()];
         String leaderSubPath = TEST_PATH + "-leader-";
         assertTrue(checkLimitEnforcedForServer(leaderServer, leaderSubPath, CreateMode.EPHEMERAL));
@@ -83,13 +104,13 @@ public class EphemeralNodeThrottlingTest extends QuorumPeerTestBase {
         long actual = (long) MetricsUtils.currentServerMetrics().get(EPHEMERAL_BYTE_LIMIT_VIOLATION_KEY);
         assertEquals(2, actual);
 
-        servers.shutDownAllServers();
+        // waitForAll(servers, ZooKeeper.States.CONNECTING);
+        // ServerMetrics.DEFAULT_METRICS_FOR_TESTS.resetAll();
     }
 
     @Test
     public void limitingSequentialEphemeralsTest() throws Exception {
         System.setProperty(EPHEMERAL_BYTE_LIMIT_KEY, Integer.toString(DEFAULT_EPHEMERALNODES_TOTAL_BYTE_LIMIT));
-        servers = LaunchServers(NUM_SERVERS);
         ZooKeeper leaderServer = servers.zk[servers.findLeader()];
         String leaderSubPath = TEST_PATH + "-leader-";
         assertTrue(checkLimitEnforcedForServer(leaderServer, leaderSubPath, CreateMode.EPHEMERAL_SEQUENTIAL));
@@ -102,7 +123,8 @@ public class EphemeralNodeThrottlingTest extends QuorumPeerTestBase {
         long actual = (long) MetricsUtils.currentServerMetrics().get(EPHEMERAL_BYTE_LIMIT_VIOLATION_KEY);
         assertEquals(2, actual);
 
-        servers.shutDownAllServers();
+        // waitForAll(servers, ZooKeeper.States.CONNECTING);
+        // ServerMetrics.DEFAULT_METRICS_FOR_TESTS.resetAll();
     }
 
     public boolean checkLimitEnforcedForServer(ZooKeeper server, String subPath, CreateMode mode) throws Exception {
@@ -146,8 +168,6 @@ public class EphemeralNodeThrottlingTest extends QuorumPeerTestBase {
 
     @Test
     public void rejectedEphemeralMetricsTest() throws Exception {
-        System.setProperty(EPHEMERAL_BYTE_LIMIT_KEY, Integer.toString(DEFAULT_EPHEMERALNODES_TOTAL_BYTE_LIMIT));
-        servers = LaunchServers(NUM_SERVERS);
         ZooKeeper leaderServer = servers.zk[servers.findLeader()];
         int expectedLimitExceededAttempts = 10;
         int i = expectedLimitExceededAttempts;
@@ -168,93 +188,7 @@ public class EphemeralNodeThrottlingTest extends QuorumPeerTestBase {
         long actual = (long) MetricsUtils.currentServerMetrics().get(EPHEMERAL_BYTE_LIMIT_VIOLATION_KEY);
         assertEquals(expectedLimitExceededAttempts, actual);
 
-        servers.shutDownAllServers();
-    }
-
-    // Tests multithreaded creates and deletes against the leader and a follower server
-    @Test
-    public void multithreadedRequestsTest() throws Exception {
-        // 50% of 1mb jute max buffer
-        int totalEphemeralNodesByteLimit = (int) (Math.pow(2d, 20d) * .5);
-        System.setProperty("zookeeper.ephemeralNodes.total.byte.limit", Integer.toString(totalEphemeralNodesByteLimit));
-
-        servers = LaunchServers(NUM_SERVERS);
-        ZooKeeper leaderServer = servers.zk[servers.findLeader()];
-        ZooKeeper followerServer = servers.zk[servers.findAnyFollower()];
-
-        runMultithreadedRequests(leaderServer);
-        runMultithreadedRequests(followerServer);
-
-        // TODO: What % delta do we want to allow here?
-        // Expensive calculation of total byte size for session ephemerals
-        long time = System.currentTimeMillis();
-        int leaderSessionEphemeralsByteSum = 0;
-        for (String nodePath : leaderServer.getEphemerals()) {
-            leaderSessionEphemeralsByteSum += BinaryOutputArchive.getSerializedStringByteSize(nodePath);
-        }
-        assertEquals(totalEphemeralNodesByteLimit, leaderSessionEphemeralsByteSum, totalEphemeralNodesByteLimit/20d);
-
-        int followerSessionEphemeralsByteSum = 0;
-        for (String nodePath : leaderServer.getEphemerals()) {
-            followerSessionEphemeralsByteSum += BinaryOutputArchive.getSerializedStringByteSize(nodePath);
-        }
-        assertEquals(totalEphemeralNodesByteLimit, followerSessionEphemeralsByteSum, totalEphemeralNodesByteLimit/20d);
-
-        System.out.println("--- total time to calculate sizes was : " + (System.currentTimeMillis() - time) + " ms -----");
-        servers.shutDownAllServers();
-    }
-
-    private void runMultithreadedRequests(ZooKeeper server) {
-        int threadPoolCount = 8;
-        int deleteRequestThreads = 2;
-        int createRequestThreads = threadPoolCount - deleteRequestThreads;
-        // Spin up threads to repeatedly send CREATE requests to server
-        ExecutorService executor = Executors.newFixedThreadPool(threadPoolCount);
-        for (int i = 0; i < createRequestThreads; i++) {
-            final int threadID = i;
-            executor.submit(() ->{
-                long startTime = System.currentTimeMillis();
-                while (System.currentTimeMillis() - startTime < 10000) {
-                    try {
-                        server.create(TEST_PATH +"_"+threadID+"_", new byte[512], ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.EPHEMERAL_SEQUENTIAL);
-                    } catch (KeeperException.TotalEphemeralLimitExceeded expectedException) {
-                        //  Ignore Ephemeral Count exceeded exception, as this is expected to occur
-                    }
-                }
-            });
-        }
-
-        // Spin up threads to repeatedly send DELETE requests to server
-        // After a 1-second sleep, this should run concurrently with the create threads, but then end before create threads
-        // so that we still have time to hit the limit and can then assert that limit was upheld correctly
-        for (int i = 0; i < deleteRequestThreads; i++) {
-            executor.submit(() -> {
-                long startTime = System.currentTimeMillis();
-                try {
-                    // Brief sleep to reduce chance that ephemeral nodes not yet created
-                    Thread.sleep(1000);
-                    while (System.currentTimeMillis() - startTime < 6000) {
-                        for (String ephemeralNode : server.getEphemerals()) {
-                            server.delete(ephemeralNode, -1);
-                        }
-                    }
-                } catch (KeeperException.TotalEphemeralLimitExceeded expectedException) {
-                    //  Ignore Ephemeral Count exceeded exception, as this is expected to occur
-                } catch (Exception e) {
-                    LOG.error("Thread encountered an exception but ignored it:\n" + e.getMessage());
-                }
-            });
-        }
-
-        executor.shutdown();
-        try {
-            if(!executor.awaitTermination(12000, TimeUnit.MILLISECONDS)) {
-                LOG.warn("Threads did not finish in the given time!");
-                executor.shutdownNow();
-            }
-        } catch (InterruptedException e) {
-            LOG.error(e.getMessage());
-            executor.shutdownNow();
-        }
+        // waitForAll(servers, ZooKeeper.States.CONNECTING);
+        // ServerMetrics.DEFAULT_METRICS_FOR_TESTS.resetAll();
     }
 }
